move helper methods to file for use in all test modules
