+++ test/lib/client_patient_test.rb
@@ -1,0 +1,6 @@
+require 'test_helper'
+
+class ClientPatientTest < Minitest::Test
+
+
+end
