+++ test/lib/client_test.rb
@@ -44,75 +44,8 @@ class ClientTest < Minitest::Test
     VCR.use_cassette('patient/new_test') do
       r = redox
       response = r.add_patient(real_patient)
-      p response, r.response
+      assert_equal 200, r.response.code.to_i
+      refute_nil response[:patient]
     end
-  end
-
-  private
-
-  def test_redox(refresh: nil, access: nil)
-    Redox::Client.new(
-      source: source,
-      destinations: destinations,
-      test: true,
-      token: access,
-      refresh_token: refresh
-    )
-  end
-
-  def redox(refresh: nil, access: nil)
-    Redox::Client.new(
-      source: real_source,
-      destinations: real_destinations,
-      test: true,
-      token: access,
-      refresh_token: refresh
-    )
-  end
-
-  def request_body_new_patient
-    {
-      Meta: {
-        DataModel: 'PatientAdmin',
-        EventType: 'NewPatient',
-        Test: true,
-        Source: source,
-        Destinations: destinations
-      },
-      Patient: patient
-    }
-  end
-
-  def source
-    {
-      Name: 'Redox Dev Tools',
-      ID: '4-5-6'
-    }
-  end
-
-  def real_source
-    redox_keys[:source_data]
-  end
-
-  def destinations
-    [
-      {
-        Name: 'Redox EMR',
-        ID: '7-8-9'
-      }
-    ]
-  end
-
-  def real_destinations
-    redox_keys[:destinations_data]
-  end
-
-  def patient
-    {
-      Identifiers: [],
-      Demographics: {
-        FirstName: 'Joe'
-      }
-    }
   end
 end
